+++ accounting/RELEASES.md
@@ -69,3 +69,13 @@ Based on:
 - [typescript v0.5.0] accounting
 ### Releases
 - [NPM v0.5.0] https://www.npmjs.com/package/@speakeasy-sdks/accounting/v/0.5.0 - accounting
+
+## 2023-12-25 01:26:05
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Accounting.yaml
+- Speakeasy CLI 1.129.1 (2.223.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.5.1] accounting
+### Releases
+- [NPM v0.5.1] https://www.npmjs.com/package/@speakeasy-sdks/accounting/v/0.5.1 - accounting